nxtrix-crm/requirements_streamlit.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
openai>=0.28.0
supabase>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
twilio>=8.0.0
scikit-learn>=1.3.0
stripe>=5.5.0